requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.7.0
asyncprawcore==2.3.0
attrs==23.1.0
backcall==0.2.0
blosc2==2.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Cython==0.29.33
decorator==5.1.1
executing==1.2.0
fastapi==0.94.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.4
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.0
numexpr==2.8.4
numpy==1.24.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
praw==7.6.1
prawcore==2.3.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pydantic==1.10.6
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pyvis==0.3.1
PyYAML==6.0
redis==4.5.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
stack-data==0.6.2
starlette==0.26.0.post1
tables==3.8.0
traitlets==5.9.0
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.14
uvicorn==0.21.0
wcwidth==0.2.6
websocket-client==1.5.1
XlsxWriter==3.0.8
yarl==1.8.2
